+++ src/main/scala/intent/runner/TestSuiteRunner.scala
@@ -57,7 +57,7 @@ class TestSuiteRunner(classLoader: ClassLoader) with
    * Until the full suite is executed a subscriber can be given to receive events during the test-run. A custom
    * subscriber is also recommended if more details than the successful result is needed.
    */
-  def runSuite(className: String, eventSubscriber: Option[Subscriber[TestCaseResult]] = None)(given ec: ExecutionContext): Future[Either[TestSuiteError, TestSuiteResult]] =
+  def runSuite(className: String, eventSubscriber: Option[Subscriber[TestCaseResult]] = None)(using ec: ExecutionContext): Future[Either[TestSuiteError, TestSuiteResult]] =
     instantiateSuite(className) match
       case Success(instance) => runTestsForSuite(instance, eventSubscriber).map(res => Right(res))
       case Failure(ex: Throwable) => Future.successful(Left(TestSuiteError(ex)))
@@ -70,7 +70,7 @@ class TestSuiteRunner(classLoader: ClassLoader) with
       case Success(instance) => Right(instance)
       case Failure(ex: Throwable) => Left(TestSuiteError(ex))
 
-  private def runTestsForSuite(suite: IntentStructure, eventSubscriber: Option[Subscriber[TestCaseResult]])(given ec: ExecutionContext): Future[TestSuiteResult] =
+  private def runTestsForSuite(suite: IntentStructure, eventSubscriber: Option[Subscriber[TestCaseResult]])(using ec: ExecutionContext): Future[TestSuiteResult] =
     // TODO: We should measure Suite time as well. Might be good to find expensive setup or scheduling problems.
     val futureTestResults = suite.allTestCases.map(tc =>
       eventSubscriber match {